requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pytest==8.3.1
httpx==0.27.2
sqlalchemy>=2.0.36
psycopg2-binary==2.9.9
redis>=5.0.0

